requirements.txt
accelerate==1.2.1
acres==0.3.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio~=3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi~=1.15.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser~=5.3.0
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
durationpy==0.9
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi~=0.103.2
fastjsonschema==2.20.0
filelock~=3.8.0
fitz==0.0.1.dev2
flatbuffers==24.3.25
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib~=1.3.2
json5==0.9.28
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.22
langchain-chroma==0.1.4
langchain-community==0.3.20
langchain-core==0.3.49
langchain-huggingface==0.1.2
langchain-openai==0.2.9
langchain-text-splitters==0.3.7
langsmith==0.1.143
llama_cpp_python==0.3.8
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe~=2.0.1
marshmallow==3.23.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx~=2.6.3
nibabel~=4.0.2
nipype~=1.8.6
notebook==7.2.2
notebook_shim==0.2.4
numpy~=1.21.6
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.5
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.11
overrides==7.7.0
packaging~=23.0
pandas~=1.3.5
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
posthog==3.7.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==6.30.2
prov==2.0.1
psutil==6.1.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic~=2.5.3
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydot~=2.0.0
Pygments~=2.17.2
PyMuPDF==1.25.5
pyparsing~=3.1.4
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pywin32==308
pywinpty==2.0.14
pyxnat==1.6.3
PyYAML~=6.0
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests~=2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich~=13.8.1
rpds-py==0.21.0
rsa==4.9
safetensors==0.5.2
scikit-learn~=1.0.2
scipy~=1.7.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.0.1
setuptools~=47.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve~=2.4.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette~=0.27.0
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
threadpoolctl~=3.1.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
traits~=6.3.2
transformers==4.48.0
typer==0.13.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3~=2.0.7
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.45.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.2
zipp~=3.8.1
passlib~=1.7.4
jose~=1.0.0
argparse~=1.4.0
pytest~=7.2.1
future~=0.18.2
pip~=20.1.1
redis~=4.4.2
exceptiongroup~=1.1.0
gunicorn~=20.0.4
itsdangerous~=1.1.0
chardet~=3.0.4